backend/requirements.txt
fastapi
uvicorn
sqlalchemy
requests
pydantic
pydantic-settings
python-dotenv
python-multipart
asyncpg
httpx
dnspython
